services/avena_video/requirements.txt
python-ffmpeg==2.0.4
nats-py==2.2.0
